--- YAHHOI_4/Assets/EnemyHP.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2bfe4fef7bda97f4bba6aa4c0520fc49
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
