+++ YAHHOI_4/Assets/Items/DoreMihara/スタンガン.png
