+++ YAHHOI_4/Assets/Items/DoreMihara/魚雷.png
