+++ YAHHOI_4/Assets/Items/EneMihara/cookies.png.meta
