+++ YAHHOI_4/Assets/Items/EneMihara/回復薬(小).png.meta
