+++ YAHHOI_4/Assets/Items/SashiMihara.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5e33ebd5443f76d4faa86fe41a47cc8f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
