+++ YAHHOI_4/Assets/Items/SashiMihara/SimoiHara.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 76f1c9737a9f65e48a6c1825acdc6a29
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
