+++ YAHHOI_4/Assets/Items/SashiMihara/SimoiHara/ToriMihara.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 168012bcead0b974695fcca3144392ef
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
