+++ YAHHOI_4/Assets/Items/SashiMihara/SimoiHara/ToriMihara/HaraMihara.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db503f5e39ec5de4d95a933d1cab815e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
