+++ YAHHOI_4/Assets/Items/SashiMihara/SimoiHara/ToriMihara/HaraMihara/SiziMihara.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 24a49c8c505951840a50403c91834da5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
