+++ YAHHOI_4/Assets/Items/SashiMihara/SimoiHara/ToriMihara/HaraMihara/SiziMihara/HotateHara.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 49240ac3f2e6aec49adcefde73494509
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
