+++ YAHHOI_4/Assets/Items/SashiMihara/SimoiHara/ToriMihara/HaraMihara/SiziMihara/HotateHara/NezuMiHara.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad7f4c2794757f647a0c8c77006f4d27
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
