+++ YAHHOI_4/Assets/Ogawa/gazou/fishS.prefab
@@ -10,6 +10,9 @@ GameObject:
   m_Component:
   - component: {fileID: 2336564056615889418}
   - component: {fileID: 3351382680276639219}
+  - component: {fileID: -3188188417380466113}
+  - component: {fileID: 2799244872330009612}
+  - component: {fileID: 6193920011629616908}
   m_Layer: 0
   m_Name: fishS
   m_TagString: Enemy
@@ -82,3 +85,114 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
+--- !u!60 &-3188188417380466113
+PolygonCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2861639284253387008}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: 0}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 2, y: 2.09}
+    newSize: {x: 2, y: 2.09}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  m_Points:
+    m_Paths:
+    - - {x: 0.74, y: -0.415}
+      - {x: 0.55, y: -0.33499998}
+      - {x: 0.35, y: -0.235}
+      - {x: 0.35, y: 0.16499999}
+      - {x: 0.44, y: 0.24499999}
+      - {x: 0.44, y: 0.505}
+      - {x: 0.28, y: 0.66499996}
+      - {x: 0.17, y: 0.66499996}
+      - {x: 0.08, y: 0.645}
+      - {x: -0.049999997, y: 0.775}
+      - {x: -0.17, y: 0.885}
+      - {x: -0.31, y: 0.935}
+      - {x: -0.59, y: 0.935}
+      - {x: -0.71, y: 0.815}
+      - {x: -0.71, y: 0.525}
+      - {x: -0.65999997, y: 0.435}
+      - {x: -0.62, y: 0.405}
+      - {x: -0.62, y: -0.155}
+      - {x: -0.58, y: -0.255}
+      - {x: -0.48999998, y: -0.525}
+      - {x: -0.42, y: -0.585}
+      - {x: -0.39, y: -0.575}
+      - {x: -0.37, y: -0.555}
+      - {x: -0.37, y: -0.49499997}
+      - {x: -0.42, y: -0.45499998}
+      - {x: -0.42, y: -0.21499999}
+      - {x: -0.16, y: -0.21499999}
+      - {x: 0.089999996, y: -0.445}
+      - {x: 0.19999999, y: -0.655}
+      - {x: 0.22999999, y: -0.755}
+      - {x: 0.32999998, y: -0.85499996}
+      - {x: 0.37, y: -0.85499996}
+      - {x: 0.53999996, y: -0.695}
+      - {x: 0.61, y: -0.675}
+      - {x: 0.81, y: -0.525}
+      - {x: 0.81, y: -0.49499997}
+    - - {x: -0.37, y: -0.565}
+      - {x: -0.39, y: -0.575}
+      - {x: -0.39, y: -0.615}
+      - {x: -0.38, y: -0.625}
+      - {x: -0.34, y: -0.625}
+      - {x: -0.32999998, y: -0.60499996}
+      - {x: -0.32999998, y: -0.585}
+      - {x: -0.34, y: -0.565}
+--- !u!114 &2799244872330009612
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2861639284253387008}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2bfe4fef7bda97f4bba6aa4c0520fc49, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  effectPrefab: {fileID: 6691482098088510567, guid: 108628da5d5d94d35b86c96992487289, type: 3}
+  destroySound: {fileID: 0}
+  enemyHP: 2
+  itemprefab: {fileID: 0}
+--- !u!61 &6193920011629616908
+BoxCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2861639284253387008}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: 0}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 2, y: 2.09}
+    newSize: {x: 2, y: 2.09}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
+  m_Size: {x: 18, y: 18}
+  m_EdgeRadius: 0
