+++ YAHHOI_4/Assets/Ogawa/scripts/Enemy/DeleteMeC.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6d8abd17b54005e489725f091443fd47
+guid: 656ac40394699cb4fb2bb69024eb9d41
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
