+++ YAHHOI_4/Assets/Ogawa/scripts/Enemy/DeleteMeF.cs
@@ -14,7 +14,8 @@ public class DeleteMeF : MonoBehaviour
     }
     private void OnTriggerEnter2D(Collider2D collision)
     {
-        if (collision.gameObject.CompareTag("Bullet"))
+        if (collision.gameObject.CompareTag("Bullet")||
+            collision.gameObject.CompareTag("Wall"))
         {
             Destroy(this.gameObject);
         }
