+++ YAHHOI_4/Assets/Ogawa/scripts/Move.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a6e5929fa8fb96b40add52ac92ee7a57
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
