+++ YAHHOI_4/Assets/Ogawa/scripts/Move/EnemyMoveV2.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: df343afd7590ad84f8b29427f7d9d9cc
+guid: 22ba9baaef0b22240892c27b9df48a27
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
