+++ YAHHOI_4/Assets/Ogawa/scripts/Move/Move_LR2.cs.meta
