+++ YAHHOI_4/Assets/Ogawa/scripts/Move/RoteA.cs.meta
